Refactor code into packages
