+++ src/main/java/seedu/clialgo/command/AddCommand.java
@@ -5,7 +5,7 @@ import java.io.File;
 
 import seedu.clialgo.Buffer;
 import seedu.clialgo.FileType;
-import seedu.clialgo.TopicManager;
+import seedu.clialgo.logic.TopicManager;
 import seedu.clialgo.storage.FileManager;
 import seedu.clialgo.Ui;
