+++ src/main/java/seedu/clialgo/command/AddNoteCommand.java
@@ -1,7 +1,7 @@
 package seedu.clialgo.command;
 
 import seedu.clialgo.Buffer;
-import seedu.clialgo.TopicManager;
+import seedu.clialgo.logic.TopicManager;
 import seedu.clialgo.Ui;
 import seedu.clialgo.storage.FileManager;
 import seedu.clialgo.file.Note;
