+++ src/test/java/seedu/clialgo/command/RemoveCommandTest.java
@@ -2,7 +2,7 @@ package seedu.clialgo.command;
 
 import org.junit.jupiter.api.Test;
 import seedu.clialgo.Buffer;
-import seedu.clialgo.TopicManager;
+import seedu.clialgo.logic.TopicManager;
 import seedu.clialgo.Ui;
 import seedu.clialgo.storage.FileManager;
