+++ src/test/java/seedu/clialgo/logic/TopicManagerTest.java
@@ -1,4 +1,4 @@
-package seedu.clialgo;
+package seedu.clialgo.logic;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -10,6 +10,8 @@ import java.util.Set;
 import org.junit.jupiter.api.Test;
 import seedu.clialgo.file.CS2040CFile;
 import seedu.clialgo.file.Note;
+import seedu.clialgo.logic.Topic;
+import seedu.clialgo.logic.TopicManager;
 import seedu.clialgo.storage.FileManagerStub;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
